[CHORE] 내가 쓴 댓글 수정 API 및 북마크 Method 변경

+++ controller/mypage/deleteMycomments.js
@@ -11,26 +11,29 @@ const deleteMycomments = async (req, res, next) => {
     }
 
     for (let i = 0; i < commentIdList.length; i++) {
-      const Post = await models.Post.findOne({
-        where: { id: Comment.postId }
-      });
-
       const Comment = await models.Comment.findOne({
         attributes: ['postId'],
         where: { id: commentIdList[i] }
       });
 
-      await models.Comment.destroy({
-        where: { id: commentIdList[i] }
+      const Post = await models.Post.findOne({
+        where: { id: Comment.postId }
       });
 
-      const num = Post.numberOfComment - 1;
-      await models.Post.update(
-        {
-          numberOfComment: num
-        },
-        { where: { id: Comment.postId } }
-      );
+      if (!Post) {
+        return res.stauts(400).send('POST_DOSENT_EXIST');
+      } else {
+        await models.Comment.destroy({
+          where: { id: commentIdList[i] }
+        });
+        const num = Post.numberOfComment - 1;
+        await models.Post.update(
+          {
+            numberOfComment: num
+          },
+          { where: { id: Comment.postId } }
+        );
+      }
     }
 
     return res.status(200).send('DELETE_MY_COMMENT_SUCCESS');
